Add files via upload
This code, which I've run on the Spyder IDE, produces a special set of random matrices. They are Haar random unitary matrices, it is possible to check whether these matrices are Haar random by acting the set of matrices in question on an initial state/vector and obtaining a set of probabilities for a certain state. The normalized frequency of probabilities will resemble a distribution, called the Porter-Thomas distribution, provided that the matrices are Haar random. 

Please consult with the document titled "V&V Presentation" for a more detailed and visual explanation, together with the code titled "haar spider".